[Chore #161986255] update tests
